pyproject.toml
[project]
name = "sched-agent"
version = "0.1.0"
description = "FastMCP + LangGraph based LLM Agent Server"
requires-python = ">=3.12"
dependencies = [
    "fastmcp",
    "langchain",
    "langgraph",
    "openai",
    "pydantic",
    "uvicorn",
    "mlflow",
    "streamlit",
    "tavily-python>=0.7.2",
    "dotenv>=0.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app" 
